+++ src/Elements/Geometry/Arc.cs
@@ -59,47 +59,13 @@ namespace Elements.Geometry
         /// Return transform on the arc at parameter u.
         /// </summary>
         /// <param name="u">A parameter between 0.0 and 1.0 on the arc.</param>
-        /// <param name="rotation">An optional rotation of the transform around its Z axis.</param>
         /// <returns>A transform with its origin at u along the curve and its Z axis tangent to the curve.</returns>
-        public override Transform TransformAt(double u, double rotation = 0.0)
+        public override Transform TransformAt(double u)
         {
             var p = PointAt(u);
             var x = (p-this.Center).Normalized();
             var y = Vector3.ZAxis;
-            return new Transform(p, x, x.Cross(y), rotation);
-        }
-
-        /// <summary>
-        /// Get a collection of transforms which represent frames along the arc.
-        /// </summary>
-        /// <param name="startSetback">The parameter offset from the start of the arc. Between 0 and 1.</param>
-        /// <param name="endSetback">The parameter offset from the end of the arc. Between 0 and 1.</param>
-        /// <param name="rotation">An optional rotation for all frames around their Z axes.</param>
-        /// <returns>A collection of transforms.</returns>
-        public override Transform[] Frames(double startSetback, double endSetback, double rotation = 0.0)
-        {
-            var div = 10;
-            var l = this.Length();
-            var step = (1.0 - startSetback - endSetback)  / div;
-            var result = new Transform[div + 1];
-            for (var i = 0; i <= div; i++)
-            {
-                Transform t;
-                if (i == 0)
-                {
-                    t = TransformAt(0.0 + startSetback, rotation);
-                }
-                else if (i == div)
-                {
-                    t = TransformAt(1.0 - endSetback, rotation);
-                }
-                else
-                {
-                    t = TransformAt(startSetback + i * step, rotation);
-                }
-                result[i] = t;
-            }
-            return result;
+            return new Transform(p, x, x.Cross(y));
         }
 
         /// <summary>
@@ -136,17 +102,47 @@ namespace Elements.Geometry
             return new Plane(this.PointAt(0.0), this.PointAt(1.0), this.Center);
         }
 
+        internal override double[] GetSampleParameters(double startSetback = 0.0, double endSetback = 0.0)
+        {
+            // Arc length calculations.
+            // var l = this.Length();
+            // var arcLength = l - startSetback - endSetback;
+            // al = (alpha * PI * r) / 180
+            // alpha = (al * 180)/(PI * r)
+            // var arcAngle = (arcLength * 360) / (2 * Math.PI * this.Radius);
+            // var parameterSpan = 1.0 - startSetback/l - endSetback/l;
+
+            // Parameter calculations.
+            var angleSpan = this.EndAngle - this.StartAngle;
+            var partialAngleSpan = angleSpan - angleSpan * startSetback - angleSpan * endSetback;
+            var parameterSpan = 1.0 - 1.0 * startSetback - 1.0 * endSetback;
+
+            // Angle span: t
+            // d = 2 * r * sin(t/2)
+            var d = MinimumChordLength;
+            var r = this.Radius;
+            var t = 2 * Math.Asin(d / (2 * r));
+            var div = (int)Math.Ceiling((DegToRad(partialAngleSpan))/t);
+
+            var parameters = new double[div+1];
+            for (var i = 0; i <= div; i++)
+            {
+                var u = startSetback + i * (parameterSpan/div);
+                parameters[i] = u;
+            }
+            return parameters;
+        }
+
         /// <summary>
         /// A list of vertices describing the arc for rendering.
         /// </summary>
         internal override IList<Vector3> RenderVertices()
         {
-            var div = 10;
-            var vertices = new Vector3[div];
-            for (var i = 0; i < div; i++)
+            var parameters = GetSampleParameters();
+            var vertices = new List<Vector3>();
+            foreach(var p in parameters)
             {
-                var t = i * (1.0/div);
-                vertices[i] = PointAt(t);
+                vertices.Add(PointAt(p));
             }
             return vertices;
         }