+++ src/Elements/Geometry/Curve.cs
@@ -8,8 +8,16 @@ namespace Elements.Geometry
     [JsonInheritanceAttribute("Elements.Geometry.Arc", typeof(Arc))]
     [JsonInheritanceAttribute("Elements.Geometry.Polyline", typeof(Polyline))]
     [JsonInheritanceAttribute("Elements.Geometry.Polygon", typeof(Polygon))]
+    [JsonInheritanceAttribute("Elements.Geometry.Bezier", typeof(Bezier))]
+    [JsonInheritanceAttribute("Elements.Geometry.Circle", typeof(Circle))]
     public abstract partial class Curve : ICurve
     {
+        /// <summary>
+        /// The minimum chord length allowed for subdivision of the curve.
+        /// A lower MinimumChordLength results in smoother curves.
+        /// </summary>
+        public static double MinimumChordLength = 0.1;
+
         /// <summary>
         /// Get the bounding box for this curve.
         /// </summary>
@@ -19,12 +27,20 @@ namespace Elements.Geometry
         /// <summary>
         /// Get a collection of transforms which represent frames along this curve.
         /// </summary>
-        /// <param name="startSetback">The offset from the start of the curve.</param>
-        /// <param name="endSetback">The offset from the end of the curve.</param>
-        /// <param name="rotation">An optional rotation of the transform around its Z axis.</param>
+        /// <param name="startSetback">The offset parameter from the start of the curve.</param>
+        /// <param name="endSetback">The offset parameter from the end of the curve.</param>
         /// <returns>A collection of transforms.</returns>
-        public abstract Transform[] Frames(double startSetback = 0, double endSetback = 0, double rotation = 0.0);
-        
+        public virtual Transform[] Frames(double startSetback = 0.0, double endSetback = 0.0)
+        {
+            var parameters = GetSampleParameters(startSetback, endSetback);
+            var transforms = new Transform[parameters.Length];
+            for (var i = 0; i < parameters.Length; i++)
+            {
+                transforms[i] = TransformAt(parameters[i]);
+            }
+            return transforms;
+        }
+
         /// <summary>
         /// Calculate the length of the curve.
         /// </summary>
@@ -42,9 +58,13 @@ namespace Elements.Geometry
         /// positive Z axis points along the curve.
         /// </summary>
         /// <param name="u">The parameter along the Line, between 0.0 and 1.0, at which to calculate the Transform.</param>
-        /// <param name="rotation">An optional rotation of the transform around its Z axis.</param>
         /// <returns>A transform.</returns>
-        public abstract Transform TransformAt(double u, double rotation = 0.0);
+        public abstract Transform TransformAt(double u);
+
+        internal virtual double[] GetSampleParameters(double startSetback = 0.0, double endSetback = 0.0)
+        {
+            return new[] { startSetback, 1.0 - endSetback };
+        }
 
         /// <summary>
         /// A list of vertices used to render the curve.