+++ src/Elements/Geometry/Profiles/ProfileServer.cs
@@ -6,7 +6,7 @@ namespace Elements.Geometry.Profiles
     /// <summary>
     /// Base class for all types which serve profiles.
     /// </summary>
-    public abstract class ProfileServer
+    public abstract class ProfileServer<T> where T: struct
     {
         /// <summary>
         /// A conversion factor from inches to meters.
@@ -16,21 +16,42 @@ namespace Elements.Geometry.Profiles
         /// <summary>
         /// The map of Profiles.
         /// </summary>
-        protected Dictionary<string, Profile> _profiles = new Dictionary<string, Profile>();
+        protected Dictionary<T, Profile> _profiles = new Dictionary<T, Profile>();
 
         /// <summary>
         /// Get a profile by name from the server.
         /// </summary>
         /// <param name="name"></param>
-        /// <returns>A Profile. Throws an exception if a profile with the specified name cannot be found.</returns>
-        /// <exception>Thrown when a Profile with the specfied name cannot be found.</exception>
+        /// <returns>A profile. Throws an exception if a profile with the specified name cannot be found.</returns>
+        /// <exception>Thrown when a profile with the specfied name cannot be found.</exception>
+        [ObsoleteAttribute("GetProfileByName will no longer be supported. Use GetProfileByType instead.")]
         public Profile GetProfileByName(string name)
         {
-            if (!this._profiles.ContainsKey(name))
+            if(!Enum.TryParse(name, true, out T result))
             {
-                throw new Exception($"The specified Wide Flange profile name, {name}, could not be found.");
+                throw new Exception($"The profile with the name, {name}, is not recognized for this profile server.");
             }
-            return this._profiles[name];
+
+            if (!this._profiles.ContainsKey(result))
+            {
+                throw new Exception($"The profile with the name, {name}, could not be found.");
+            }
+            return this._profiles[result];
+        }
+
+        /// <summary>
+        /// Get a profile by type enumeration from the server.
+        /// </summary>
+        /// <param name="type">The enumerated type of the profile.</param>
+        /// <returns>A profile. Throws an exception if a profile with the specified name cannot be found.</returns>
+        /// <exception>Thrown when a profile of the specfied type cannot be found.</exception>
+        public Profile GetProfileByType(T type)
+        {
+            if (!this._profiles.ContainsKey(type))
+            {
+                throw new Exception($"The profile with the name, {type.ToString()}, could not be found.");
+            }
+            return this._profiles[type];
         }
 
         /// <summary>