+++ test/Elements.Tests/Examples/ModelCurveExample.cs
@@ -13,7 +13,7 @@ namespace Elements.Tests.Examples
             // <example>
             // Create some curves for your model curves.
             var line = new Line(Vector3.Origin, new Vector3(5,5,5));
-            var arc = new Arc(new Vector3(0,0,0), 2.0, 45.0, 135.0);
+            var arc = new Arc(Vector3.Origin, 2.0, 45.0, 135.0);
             var pline = Polygon.L(2, 2, 0.5);
 
             // Create model curves from the curves.