+++ test/Elements.Tests/SolidTests.cs
@@ -112,7 +112,7 @@ namespace Elements.Tests
         [Fact]
         public void SweptSolidPolyline()
         {
-            var profile = WideFlangeProfileServer.Instance.GetProfileByName("W44x335");
+            var profile = WideFlangeProfileServer.Instance.GetProfileByType(WideFlangeProfileType.W10x100);
             var path = new Polyline(new []{new Vector3(0,0), new Vector3(0,2), new Vector3(0,3,1), new Vector3(0,5,1)});
             var solid = Solid.SweepFaceAlongCurve(profile.Perimeter, null, path);
             foreach(var e in solid.Edges.Values)
@@ -127,7 +127,7 @@ namespace Elements.Tests
         [Fact]
         public void SweptSolidArc()
         {
-            var profile = WideFlangeProfileServer.Instance.GetProfileByName("W44x335");
+            var profile = WideFlangeProfileServer.Instance.GetProfileByType(WideFlangeProfileType.W10x100);
             var path = new Arc(Vector3.Origin, 5, 0, 90);
             var solid = Solid.SweepFaceAlongCurve(profile.Perimeter, null, path);
             foreach(var e in solid.Edges.Values)
@@ -142,7 +142,7 @@ namespace Elements.Tests
         [Fact]
         public void SweptSolidPolygon()
         {
-            var profile = WideFlangeProfileServer.Instance.GetProfileByName("W44x335");
+            var profile = WideFlangeProfileServer.Instance.GetProfileByType(WideFlangeProfileType.W10x100);
             var path = Polygon.Ngon(12, 5);
             var solid = Solid.SweepFaceAlongCurve(profile.Perimeter, null, path);
             foreach(var e in solid.Edges.Values)